+++ apps/openmw/mwworld/containerstore.cpp
@@ -18,6 +18,7 @@
 */
 
 #include <components/debug/debuglog.hpp>
+#include <components/misc/stringops.hpp>
 #include <components/esm/inventorystate.hpp>
 
 #include "../mwbase/environment.hpp"